Touch events for instr and sample editors
(also fixes some interactions with UI components like scroll bars,
maybe)

There are two parts to this:

  1. Manually calling mouseButton{Down,Up}Handler with SDL_BUTTON_LEFT
     for touch events.
  2. Reading the mouse X and Y from any touch which is present in
     readMouseXY.

The first part ensures the ft2clone logic for processing mouse up/down
events is triggered for touch events, and the second part ensures that
`mouse.x` and `mouse.y` are updated whenever there is an active touch
(not just when a finger up/down event occurs.

I thought maybe this would mean we could get rid of the hack where we
set `mouse.x` and `mouse.y` whenever a touch event occurs, however this
is still needed - I guess because when a touch event triggers, the event
handling code runs before the `readMouseXY` in the main loop gets a
chance to update `mouse.x` and `mouse.y` from the active touch.

Limitations:

  1. All touches are left mouse button touches.
  2. If multiple touch devices are present and have touches then the
     last device index is the one whose touch will be used.